edit model, repository and service

+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/repository/AccountRepository.java
@@ -10,11 +10,11 @@ public interface AccountRepository {
 
     void insert(String user, String pass, boolean isSell, boolean isAdmin, String email);
 
-    void deleteByID(long id);
+    void deleteByID(int id);
 
-    void edit(String username, String password, boolean isSell, boolean isAdmin, String email, long uID);
+    void edit(String username, String password, boolean isSell, boolean isAdmin, String email, int uID);
 
-    void updateProfile(String username, String password, String email, long uID);
+    void updateProfile(String username, String password, String email, int uID);
 
     void signIn(String user, String pass);
 
@@ -24,7 +24,7 @@ public interface AccountRepository {
 
     List<Account> getAll();
 
-    Account getById(long id);
+    Account getById(int id);
 
     List<Account> getListByPage(List<Account> list, int start, int end);
 }
