+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/repository/InvoiceRepository.java
@@ -5,9 +5,9 @@ import com.HE180030.model.Invoice;
 import java.util.List;
 
 public interface InvoiceRepository {
-    void insert(long accountID, double totalPrice, String context, int phone, String delivery, String name, String typePay);
+    void insert(int accountID, double totalPrice, String context, int phone, String delivery, String name, String typePay);
 
-    void deleteByAccountID(long id);
+    void deleteByAccountID(int id);
 
-    List<Invoice> getAllInvoiceByID(long id);
+    List<Invoice> getAllInvoiceByID(int id);
 }
