+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/repository/QuantitiesSoldRepository.java
@@ -1,5 +1,5 @@
 package com.HE180030.repository;
 
 public interface QuantitiesSoldRepository {
-    void deleteByProductID(long productID);
+    void deleteByProductID(int productID);
 }
