+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/repository/TotalSalesTargetRepository.java
@@ -1,5 +1,5 @@
 package com.HE180030.repository;
 
 public interface TotalSalesTargetRepository {
-    void deleteByAccountID(long id);
+    void deleteByAccountID(int id);
 }
