+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/repository/impl/QuantitiesSoldRepositoryImpl.java
@@ -18,7 +18,7 @@ public class QuantitiesSoldRepositoryImpl implements QuantitiesSoldRepository {
     }
 
     @Override
-    public void deleteByProductID(long productID) {
+    public void deleteByProductID(int productID) {
         sessionFactory.getCurrentSession()
                 .createQuery("delete from QuantitesSold q where q.product.id = :productID")
                 .setParameter("productID", productID)
