+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/service/CategoryService.java
@@ -9,5 +9,5 @@ public interface CategoryService {
 
     CategoryDTO getCategoryDTOByName(String name);
 
-    void insertCategoryDTO(long categoryId, String name);
+    void insertCategoryDTO(int categoryId, String name);
 }
