+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/service/InvoiceService.java
@@ -5,9 +5,9 @@ import com.HE180030.dto.InvoiceDTO;
 import java.util.List;
 
 public interface InvoiceService {
-    void insertInvoiceDTO(long accountID, double totalPrice, String context, int phone, String delivery, String name, String typePay);
+    void insertInvoiceDTO(int accountID, double totalPrice, String context, int phone, String delivery, String name, String typePay);
 
-    void deleteInvoiceDTOByAccountID(long id);
+    void deleteInvoiceDTOByAccountID(int id);
 
-    List<InvoiceDTO> getAllInvoiceByID(long id);
+    List<InvoiceDTO> getAllInvoiceByID(int id);
 }
