+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/service/QuantitiesSoldService.java
@@ -1,5 +1,5 @@
 package com.HE180030.service;
 
 public interface QuantitiesSoldService {
-    void deleteQuantitiesSoldDTOByProductID(long productID);
+    void deleteQuantitiesSoldDTOByProductID(int productID);
 }
