+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/service/TotalSalesTargetService.java
@@ -1,5 +1,5 @@
 package com.HE180030.service;
 
 public interface TotalSalesTargetService {
-    void deleteTotalSalesSpendDTOByAccountID(long id);
+    void deleteTotalSalesSpendDTOByAccountID(int id);
 }
