+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/service/impl/CategoryServiceImpl.java
@@ -27,7 +27,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void insertCategoryDTO(long categoryId, String name) {
+    public void insertCategoryDTO(int categoryId, String name) {
 
     }
 }
