+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/service/impl/QuantitiesSoldServiceImpl.java
@@ -13,7 +13,7 @@ public class QuantitiesSoldServiceImpl implements QuantitiesSoldService {
     }
 
     @Override
-    public void deleteQuantitiesSoldDTOByProductID(long productID) {
+    public void deleteQuantitiesSoldDTOByProductID(int productID) {
 
     }
 }
