+++ Maven webapp/Shoes_Onlineshop/src/main/java/com/HE180030/service/impl/TotalSalesTargetServiceImpl.java
@@ -13,7 +13,7 @@ public class TotalSalesTargetServiceImpl implements TotalSalesTargetService {
     }
 
     @Override
-    public void deleteTotalSalesSpendDTOByAccountID(long id) {
+    public void deleteTotalSalesSpendDTOByAccountID(int id) {
 
     }
 }
